User local server for public_url_spec and refactor
